backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "backend"
version = "0.1.0"
description = ""
authors = ["Lucas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10"
Django = "^5.0"
mysqlclient = "^2.2.4"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"